Fixed metadata setting of audio bitrate (long vs. int)

+++ MediaPortal/Source/Extensions/MetadataExtractors/Tve3RecordingMetadataExtractor/RadioRecordingMetadataExtractor.cs
@@ -95,7 +95,7 @@ namespace MediaPortal.Extensions.MetadataExtractors
 
           var audioBitrate = mediaInfo.GetAudioBitrate(0);
           if (audioBitrate.HasValue)
-            MediaItemAspect.SetAttribute(extractedAspectData, AudioAspect.ATTR_BITRATE, audioBitrate.Value / 1024); // We store kbit/s;
+            MediaItemAspect.SetAttribute(extractedAspectData, AudioAspect.ATTR_BITRATE, (int)(audioBitrate.Value / 1024)); // We store kbit/s;
 
           MediaItemAspect.SetAttribute(extractedAspectData, AudioAspect.ATTR_ENCODING, mediaInfo.GetAudioCodec(0));
           // MediaInfo returns milliseconds, we need seconds
